Fixup

+++ Modules/_ctypes/ctypes.h
@@ -141,7 +141,7 @@ extern struct fielddesc *_ctypes_get_fielddesc(const char *fmt);
 
 extern PyObject *
 PyCField_FromDesc(PyObject *desc, Py_ssize_t index,
-                Py_ssize_t *pfield_size, int bitsize, int *pbitofs,
+                Py_ssize_t *pfield_size, int bitsize, Py_ssize_t *pbitofs,
                 Py_ssize_t *psize, Py_ssize_t *poffset, Py_ssize_t *palign,
                 int pack, int is_big_endian);
